Make brush tool extend LayerGroupPointer
